feat!: move tests and helpers into  (#459)
* test: move into `test` directory

* test: update import paths

* test: update jest config

* refactor: move test helpers

* test: move jest config into dedicated file

* docs: add changelog entry

* chore: remove unneeded eslint ignore
