+++ .eslintignore
@@ -1,4 +1,3 @@
 lib/*
 node_modules/*
 vendor/*
-**/__tests__/**/*
