+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  roots: ["<rootDir>/test"],
+  testPathIgnorePatterns: [
+    "/__fixtures__/",
+    "/__utils__/"
+  ]
+}
