+++ test/package/config.test.js
@@ -1,12 +1,12 @@
 /* global test expect, describe */
 
-const { chdirTestApp, resetEnv } = require('../utils/helpers')
+const { chdirTestApp, resetEnv } = require('../helpers')
 const { resolve } = require('path')
 
 const rootPath = process.cwd()
 chdirTestApp()
 
-const config = require('../config')
+const config = require('../../package/config')
 
 describe('Config', () => {
   beforeEach(() => jest.resetModules() && resetEnv())
@@ -14,14 +14,14 @@ describe('Config', () => {
 
   test('public path', () => {
     process.env.RAILS_ENV = 'development'
-    const config = require('../config')
+    const config = require('../../package/config')
     expect(config.publicPath).toEqual('/packs/')
   })
 
   test('public path with asset host', () => {
     process.env.RAILS_ENV = 'development'
     process.env.SHAKAPACKER_ASSET_HOST = 'http://foo.com/'
-    const config = require('../config')
+    const config = require('../../package/config')
     expect(config.publicPath).toEqual('http://foo.com/packs/')
   })
 
@@ -40,7 +40,7 @@ describe('Config', () => {
 
   test('should allow overriding manifestPath', () => {
     process.env.SHAKAPACKER_CONFIG = 'config/shakapacker_manifest_path.yml'
-    const config = require('../config')
+    const config = require('../../package/config')
     expect(config.manifestPath).toEqual(resolve('app/javascript/manifest.json'))
   })
 })
