+++ test/package/dev_server.test.js
@@ -1,6 +1,6 @@
 /* global test expect, describe */
 
-const { chdirTestApp } = require('../utils/helpers')
+const { chdirTestApp } = require('../helpers')
 
 const rootPath = process.cwd()
 chdirTestApp()
@@ -16,7 +16,7 @@ describe('DevServer', () => {
     process.env.SHAKAPACKER_DEV_SERVER_PORT = 5000
     process.env.SHAKAPACKER_DEV_SERVER_DISABLE_HOST_CHECK = false
 
-    const devServer = require('../dev_server')
+    const devServer = require('../../package/dev_server')
     expect(devServer).toBeDefined()
     expect(devServer.host).toEqual('0.0.0.0')
     expect(devServer.port).toEqual('5000')
@@ -24,7 +24,7 @@ describe('DevServer', () => {
   })
 
   test('with custom env prefix', () => {
-    const config = require('../config')
+    const config = require('../../package/config')
     config.dev_server.env_prefix = 'TEST_SHAKAPACKER_DEV_SERVER'
 
     process.env.NODE_ENV = 'development'
@@ -32,7 +32,7 @@ describe('DevServer', () => {
     process.env.TEST_SHAKAPACKER_DEV_SERVER_HOST = '0.0.0.0'
     process.env.TEST_SHAKAPACKER_DEV_SERVER_PORT = 5000
 
-    const devServer = require('../dev_server')
+    const devServer = require('../../package/dev_server')
     expect(devServer).toBeDefined()
     expect(devServer.host).toEqual('0.0.0.0')
     expect(devServer.port).toEqual('5000')
@@ -41,6 +41,6 @@ describe('DevServer', () => {
   test('with NODE_ENV and RAILS_ENV set to production', () => {
     process.env.RAILS_ENV = 'production'
     process.env.NODE_ENV = 'production'
-    expect(require('../dev_server')).toEqual({})
+    expect(require('../../package/dev_server')).toEqual({})
   })
 })
