+++ test/package/environments/development.test.js
@@ -1,6 +1,6 @@
 /* global test expect, describe, afterAll, beforeEach */
 
-const { chdirTestApp, resetEnv } = require('../../utils/helpers')
+const { chdirTestApp, resetEnv } = require('../../helpers')
 const rootPath = process.cwd()
 chdirTestApp()
 
@@ -14,9 +14,9 @@ describe('Development specific config', () => {
 
   describe('with config.useContentHash = true', () => {
     test('sets filename to use contentHash', () => {
-      const config = require("../../config");
+      const config = require("../../../package/config");
       config.useContentHash = true
-      const environmentConfig = require('../development')
+      const environmentConfig = require('../../../package/environments/development')
 
       expect(environmentConfig.output.filename).toEqual('js/[name]-[contenthash].js')
       expect(environmentConfig.output.chunkFilename).toEqual(
@@ -27,9 +27,9 @@ describe('Development specific config', () => {
 
   describe('with config.useContentHash = false', () => {
     test('sets filename without using contentHash', () => {
-      const config = require("../../config");
+      const config = require("../../../package/config");
       config.useContentHash = false
-      const environmentConfig = require('../development')
+      const environmentConfig = require('../../../package/environments/development')
 
       expect(environmentConfig.output.filename).toEqual('js/[name].js')
       expect(environmentConfig.output.chunkFilename).toEqual(
@@ -40,9 +40,9 @@ describe('Development specific config', () => {
 
   describe('with unset config.useContentHash', () => {
     test('sets filename without using contentHash', () => {
-      const config = require("../../config");
+      const config = require("../../../package/config");
       delete config.useContentHash
-      const environmentConfig = require('../development')
+      const environmentConfig = require('../../../package/environments/development')
 
       expect(environmentConfig.output.filename).toEqual('js/[name].js')
       expect(environmentConfig.output.chunkFilename).toEqual(
