+++ test/package/environments/production.test.js
@@ -1,6 +1,6 @@
 /* global test expect, describe, afterAll, beforeEach */
 
-const { chdirTestApp, resetEnv } = require('../../utils/helpers')
+const { chdirTestApp, resetEnv } = require('../../helpers')
 const rootPath = process.cwd()
 chdirTestApp()
 
@@ -14,9 +14,9 @@ describe('Production specific config', () => {
 
   describe('with config.useContentHash = true', () => {
     test('sets filename to use contentHash', () => {
-      const config = require("../../config");
+      const config = require("../../../package/config");
       config.useContentHash = true
-      const environmentConfig = require('../production')
+      const environmentConfig = require('../../../package/environments/production')
 
       expect(environmentConfig.output.filename).toEqual('js/[name]-[contenthash].js')
       expect(environmentConfig.output.chunkFilename).toEqual(
@@ -26,9 +26,9 @@ describe('Production specific config', () => {
 
     test("doesn't shows any warning message", () => {
       const consoleWarnSpy = jest.spyOn(console, 'warn');
-      const config = require("../../config");
+      const config = require("../../../package/config");
       config.useContentHash = true
-      const environmentConfig = require('../production')
+      const environmentConfig = require('../../../package/environments/production')
 
       expect(consoleWarnSpy).not.toHaveBeenCalledWith(
         expect.stringMatching(/Setting 'useContentHash' to 'false' in the production environment/)
@@ -40,9 +40,9 @@ describe('Production specific config', () => {
 
   describe('with config.useContentHash = false', () => {
     test('sets filename to use contentHash', () => {
-      const config = require("../../config");
+      const config = require("../../../package/config");
       config.useContentHash = false
-      const environmentConfig = require('../production')
+      const environmentConfig = require('../../../package/environments/production')
 
       expect(environmentConfig.output.filename).toEqual('js/[name]-[contenthash].js')
       expect(environmentConfig.output.chunkFilename).toEqual(
@@ -52,9 +52,9 @@ describe('Production specific config', () => {
 
     test('shows a warning message', () => {
       const consoleWarnSpy = jest.spyOn(console, 'warn');
-      const config = require("../../config");
+      const config = require("../../../package/config");
       config.useContentHash = false
-      const environmentConfig = require('../production')
+      const environmentConfig = require('../../../package/environments/production')
 
       expect(consoleWarnSpy).toHaveBeenCalledWith(
         expect.stringMatching(/Setting 'useContentHash' to 'false' in the production environment/)
@@ -66,9 +66,9 @@ describe('Production specific config', () => {
 
   describe('with unset config.useContentHash', () => {
     test('sets filename to use contentHash', () => {
-      const config = require("../../config");
+      const config = require("../../../package/config");
       delete config.useContentHash
-      const environmentConfig = require('../production')
+      const environmentConfig = require('../../../package/environments/production')
 
       expect(environmentConfig.output.filename).toEqual('js/[name]-[contenthash].js')
       expect(environmentConfig.output.chunkFilename).toEqual(
@@ -78,9 +78,9 @@ describe('Production specific config', () => {
 
     test("doesn't shows any warning message", () => {
       const consoleWarnSpy = jest.spyOn(console, 'warn');
-      const config = require("../../config");
+      const config = require("../../../package/config");
       delete config.useContentHash
-      const environmentConfig = require('../production')
+      const environmentConfig = require('../../../package/environments/production')
 
       expect(consoleWarnSpy).not.toHaveBeenCalledWith(
         expect.stringMatching(/Setting 'useContentHash' to 'false' in the production environment/)
